Complete Weather page refactor, changed API, Model, lots of stuff. Also cleaned up MatrixDisplay classes, by hiding most of the API from main.cpp, only making visible drawXXXPage() APIs

+++ src/util.h
@@ -2,7 +2,7 @@
 #define UTIL_H
 #include <Model.h>
 #include <RouteGroupData.h>
-int currentHourOfDay();
+uint8_t currentHourOfDay();
 void currentDateFull(char* buffer, size_t bufferSize);
 void currentDateShort(char* buffer, size_t bufferSize);
 void currentHourMinute(char* buffer, size_t bufferSize);
